Updated scriptage for 2021 (Cycle7).
git-svn-id: https://svn.atnf.csiro.au/difx@10263 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ sites/Haystack/ehtc/est_manual_phases.py
@@ -13,6 +13,7 @@ import os
 import re
 import subprocess
 import sys
+import time
 import textwrap
 
 def parseOptions():
@@ -30,11 +31,15 @@ def parseOptions():
     Typical usage requires just the name of the control file
     to build, the root file to use, and the list of stations:
 
-    est_manual_phases.py -r 3597/No0049/3C279.zlwmcz -c sample.conf -s A,...
+      est_manual_phases.py -r 3597/No../3C279.... -c old.conf -s A,...
+
+    To create a 'starter' control file, use the prune (-p) option:
+
+      est_manual_phases.py -r ... -c new.conf -s ... -p
     ''')
     use = '%(prog)s [options]\n'
     use += '  Version '
-    use += '$Id: est_manual_phases.py 3246 2021-07-13 17:56:08Z gbc $'
+    use += '$Id: est_manual_phases.py 3388 2021-10-15 17:34:30Z gbc $'
     parser = argparse.ArgumentParser(epilog=epi, description=des, usage=use,
         formatter_class=argparse.RawDescriptionHelpFormatter)
     required = parser.add_argument_group('required options')
@@ -65,7 +70,8 @@ def parseOptions():
         help='Nuke existing control file and start from scratch')
     flaggers.add_argument('-X', '--mixed', dest='mixed',
         action='store_true', default=False,
-        help='Fall back on the LMT/ALMA mixed first strategy')
+        help='Fall back on the (LMT)/ALMA first-station'
+        '-is-mixed-pol first strategy')
     #
     dbugging.add_argument('-d', '--dry', dest='dry',
         action='store_true', default=False,
@@ -85,11 +91,23 @@ def parseOptions():
         metavar='INT', default='4',
         help='Maximum number of iterations of -q sequence')
     optional.add_argument('-t', '--tolerance', dest='tolerance',
-        metavar='FLOAT', default=0.000001, type=float,
+        metavar='FLOAT', default=0.000010, type=float,
         help='Continue until sbd/mbd are smaller than this')
     optional.add_argument('-a', '--additional', dest='additional',
         action='store_true', default=False,
         help='If set, just do the phase/delay on the site list.')
+#   optional.add_argument('-k', '--keepers', dest='keepers',
+#       metavar='RE-LIST', default='',
+#       help='comma-sep list of RE-patterns from control file'
+#       ' to preserve in iterations of control file.  The parser for'
+#       ' the control file is rather primitive and looks for the various'
+#       ' control keywords at the start of a line.  And it is expecting'
+#       ' if station .. phases and delays to be under its control.')
+    optional.add_argument('-A', '--alma', dest='alma',
+        metavar='CHAR', default='A',
+        help='The letter of the "ALMA" type anchor station which'
+        ' must be first in the list.  "A" is the default, but you'
+        ' can change to some other station')
     #
     whatelse.add_argument('arguments', nargs='*',
         help='Any remaining command-line arguments are treated '
@@ -107,12 +125,23 @@ def parseOptions():
     o.cf_default = '''
     pc_mode manual            * manual phase required
     weak_channel 0.0          * shut up about code G
-    optimize_closure true     * expected to be better
-    mbd_anchor sbd            * model isn't too good
-    gen_cf_record true        * for debugging @
+    optimize_closure true     * use mbd not sbd for delay
+    mbd_anchor sbd            * force mbd near sbd
     sb_win -0.10     0.10     * SBD search window
     mb_win -0.008    0.008    * MBD search window
     dr_win -0.000001 0.000001 * Delay Rate window
+    if station A mount_type cassegrain  * 2021 info
+    if station G mount_type cassegrain
+    if station J mount_type cassegrain
+    if station K mount_type cassegrain
+    if station L mount_type cassegrain
+    if station M mount_type cassegrain
+    if station N mount_type cassegrain
+    if station P mount_type nasmythleft
+    if station S mount_type cassegrain
+    if station X mount_type nasmythright
+    if station Y mount_type cassegrain
+    if station Z mount_type nasmythright
     '''
     return o
 
@@ -145,6 +174,16 @@ def saveOrigControl(o):
         createNewControl(o)
     else:
         os.system('cp -p %s %s' % (o.control, o.control + '.orig'))
+
+def copyControl(o):
+    '''
+    Make a copy of the control file for debugging
+    '''
+    if not o.very: return
+    o.counter = o.counter + 1
+    cmd = 'cp -p %s %s.%03d' % (o.control, o.control, o.counter)
+    os.system(cmd)
+    print('# ' + cmd)
 
 def createFile(cfile, directives):
     '''
@@ -199,7 +238,9 @@ def executeFFact(o):
     nf = open(o.control + '.temp', 'w') 
     for line in cf.readlines(): nf.write(line)
     cf.close()
-    nf.write('\n* executeFFact\n* %s\n' % o.cmd)
+    try: etime = "%.3f"%(time.clock_gettime(time.CLOCK_REALTIME) - o.rtstart)
+    except: etime = "na"
+    nf.write('\n* executeFFact %s\n* %s\n' % (etime, o.cmd))
     converged = False
     ind = '      '
     for lineb in p.stdout:
@@ -209,11 +250,16 @@ def executeFFact(o):
             sbd = float(dly.group(1))
             mbd = float(dly.group(2))
             frr = float(dly.group(3))
-            if abs(sbd) < o.tolerance and abs(mbd) < o.tolerance:
-                converged = True
+            if (abs(sbd-o.lsbd) < o.tolerance and
+                abs(mbd-o.lmbd) < o.tolerance):
+                    converged = True
             if converged: status = '(is  converged)'
             else:         status = '(not converged)'
-            if o.verb: print(ind, line.rstrip(), sbd, mbd, frr, status)
+            if o.verb:
+                print(ind, line,
+                      ind, '* ', etime, 's', sbd, o.lsbd, mbd, o.lmbd, status)
+            o.lsbd = sbd
+            o.lmbd = mbd
         else:
             if o.very: print(ind, line, end=' ')
         # protect control file from error messages
@@ -223,6 +269,7 @@ def executeFFact(o):
         nf.write(line)
     p.wait()
     nf.close()
+    copyControl(o)
     os.unlink(o.control)
     os.rename(o.control + '.temp', o.control)
     return converged
@@ -251,8 +298,10 @@ def genPhaseDelay(o, ref, rem, ref_pol, rem_pol, sgn):
         if o.verb: print('  Missing datafile %s, skipping...' % cor)
         return 0
     if o.verb: print('  Using datafile %s' % cor)
+    o.lsbd = 0.0
+    o.lmbd = 0.0
     for ite in range(int(o.max)):
-        if o.very: print("# Iteration %d/%d" % ite/int(o.max))
+        if o.very: print("# Iteration %d/%d" % (ite,int(o.max)))
         for seq in o.sequence.split(','):
             o.cmd = '%s -t -c %s -b %s%s -P %s%s %s set est_pc_manual %d' % (
                 'fourfit', o.control, ref, rem, ref_pol, rem_pol,
@@ -281,12 +330,15 @@ def genPhaseOffset(o, ref, rem, ref_pol, rem_pol, sgn):
     if o.verb: print('  Using datafile %s' % cor)
     o.cmd = '%s -t -c %s -b %s%s -P %s%s %s set est_pc_manual %d' % (
         'fourfit', o.control, ref, rem, ref_pol, rem_pol, o.rootfile, sgn*64)
+    o.lsbd = 0.0
+    o.lmbd = 0.0
     if o.dry: converged = executeFFdry(o)
     else:     converged = executeFFact(o)
-    if converged:
-        if o.verb: print('    Converged')
-        return 1
-    return 0
+#   if converged:
+#       if o.verb: print('    Converged')
+#       return 1
+#   return 0
+    return 1
 
 def doTheWorkMix(o):
     '''
@@ -309,11 +361,11 @@ def doTheWorkMix(o):
     ok = 0
     eok = 4 + 2 * len(sites)
     if not o.additional:
-        if mixed != 'A':
-            raise Exception('ALMA is expected to be first in the list of sites')
+        if mixed != o.alma: raise Exception(
+            'ALMA is expected to be first in the list of sites; use\n'
+            'the -A option to have the first station play that role')
         fixed = sites.pop(0)
-        if fixed != 'L':
-            raise Exception('LMT is expected to be second in the list of sites')
+        # LMT is a great site for this, however...
         ok += genPhaseDelay(o, mixed, fixed, 'R', 'L', 1)  # step a.
         ok += genPhaseDelay(o, mixed, fixed, 'L', 'L', 1)  # step b.
         ok += genPhaseDelay(o, mixed, fixed, 'L', 'R', -1)  # step c.
@@ -325,7 +377,7 @@ def doTheWorkMix(o):
         print('The %d of %d steps were completed properly'%(ok,eok))
     else:
         print('Only %d of %d steps were completed properly'%(ok,eok))
-        #print('\n'.join(steps))
+        if o.very: print('\n'.join(steps))
 
 def doTheWorkAok(o):
     '''
@@ -334,22 +386,23 @@ def doTheWorkAok(o):
     0. build a site id mk4 id mapping list
     '''
     steps = [
-        'a. If the first site is ALMA, assume done properly: phases/delays => 0\n' +
-        '   This is the default, so there is no work. (pop)',
+        'a. If the first site is ALMA, assume it wasdone properly:\n' +
+        '   set its phases/delays => 0. This is the default, so there\n' +
+        '   is no work to this step. (pop)',
         'b. for every additional station/pol, use\n' +
         '   Ax ?? to generate x R/L phase & delay (rem)',
-        'c. If the first site is not ALMA, assume done properly; phases/delays\n' +
-        '   are not zero, but are consistent with ALMA and fourfit should be\n' +
-        '   generating numbers consistent with that.',
-        'd. for every additional station/pol, proceed to generate phases/delays\n' +
-        '   as normally.  However, it may be necessary to',
+        'c. If the first site is not ALMA, assume done properly;\n' +
+        '   phases/delays are not zero, but are consistent with ALMA\n' +
+        '   and fourfit should be generating numbers consistent with that.',
+        'd. for every additional station/pol, proceed to generate\n' +
+        '   phases/delays as normally.  However, it may be necessary to\n',
         'e. swap baselines to find the proper ref/rem combination.' ]
     doStarters(o)
     sites = o.sites.split(',')
     mixed = sites.pop(0)
     ok = 0
     eok = 2 * len(sites)
-    if mixed == 'A':
+    if mixed == o.alma:
         for other in sites:
             ok += genPhaseDelay(o, mixed, other, 'R', 'R', -1)  # step a.
             ok += genPhaseDelay(o, mixed, other, 'L', 'L', -1)  # step a.
@@ -357,7 +410,7 @@ def doTheWorkAok(o):
             print('The %d of %d steps were completed properly'%(ok,eok))
         else:
             print('Only %d of %d steps were completed properly'%(ok,eok))
-            #print('\n'.join(steps))
+            if o.very: print('\n'.join(steps))
     else:
         # trust the first station
         for other in sites:
@@ -372,28 +425,52 @@ def doTheWorkAok(o):
             print('The %d of %d steps were completed properly'%(ok,eok))
         else:
             print('Only %d of %d steps were completed properly'%(ok,eok))
-            #print('\n'.join(steps))
+            if o.very: print('\n'.join(steps))
 
 def pruneCF(o):
     '''
     Walk through the control file and prune it.
     '''
-    os.rename(o.control, o.control + '.full')
+    # os.rename(o.control, o.control + '.full')
+    if os.path.exists(o.control):
+        os.rename(o.control, o.control + '.full')
+    else:
+        createFile(o.control + '.full', o.cf_default)
     cf = open(o.control + '.full', 'r')
     nf = open(o.control, 'w')
     iftxt = ''
     site = None
     what = None
-    if_re = re.compile(r'^if station (.)')
+    # path to saving random stuff
+#   misc = [
+#       re.compile(r'^if station (.) mount_type'),
+#       re.compile(r'^ref_freq'),
+#       re.compile(r'^start'),
+#       re.compile(r'^stop'),
+#   ]
+#   for pat in o.keepers.split(','):
+#       misc.append(re.compile(pat))
+    # and the standard list we require
+    if_re = re.compile(r'^if station (.)$')
     do_re = re.compile(r'.*delay_offs_(.)')
     ph_re = re.compile(r'.*pc_phases_(.)')
     po_re = re.compile(r'.*pc_phase_offset_(.)')
+    # ok, run through the cf by line
     for lineo in cf.readlines():
         line = lineo.lstrip()
         if len(line) == 0: continue
         if line[0] == '*': continue
         if line[0] == '\n': continue
         ls = lineo.rstrip()
+#       mfor = False
+#       for m in misc:
+#           hit = m.search(ls)
+#           if hit:
+#               nf.write(line)
+#               mfor = True
+#               break # so we can continue
+#       if mfor:
+#           continue # to next line
         hit = if_re.search(ls)
         if hit:
             if site and what:
@@ -463,6 +540,9 @@ def dumpOpts(o):
 # main entry point
 if __name__ == '__main__':
     o = parseOptions()
+    o.counter = 0
+    try:    o.rtstart = time.clock_gettime(time.CLOCK_REALTIME)
+    except: o.rtstart = 'na'
     if o.very:
         o.verb = True
     if o.defaults:
@@ -473,10 +553,10 @@ if __name__ == '__main__':
         sys.exit(0)
     try:
         if o.mixed:
-            print('Assuming ALMA is mixed (-X option used).')
+            print('Assuming ALMA is mixed pol (-X option used).')
             doTheWorkMix(o)
         else:
-            print('Assuming ALMA is fixed (-X option not used).')
+            print('Assuming ALMA is fixed circ (-X option not used).')
             doTheWorkAok(o)
         pruneCF(o)
     except KeyboardInterrupt:
